ElProceso/edustream/SRC/backend/requirements.txt

fastapi
uvicorn[standard]
pymongo
python-jose
PyJWT
python-dotenv